+++ src/ConsumeResult/KafkaPartitionEof.php
@@ -4,11 +4,16 @@ declare(strict_types=1);
 
 namespace Anktx\Kafka\Client\ConsumeResult;
 
-final readonly class KafkaPartitionEof
+final readonly class KafkaPartitionEof implements ConsumeResult
 {
     public function __construct(
         public string $topic,
         public int $partition,
         public int $offset,
     ) {}
+
+    public function kind(): ConsumeResultKind
+    {
+        return ConsumeResultKind::PartitionEof;
+    }
 }